+++ jsolex/src/main/resources/whats-new.md
@@ -20,6 +20,7 @@
 - Make `angleP`, `b0`, ... parameters available in the `[[batch]]` section of scripts. A average value is computed from the images in the batch, and can be used in the script for convenience.
 - Made it possible to select an unsharp mask as the sharpness method in the image enhancement settings
 - Add corona E extraction sample script
+- Display pixel shift in the technical card
 
 ### 3.2.2